requirements.txt
requests
openai
python-dotenv
pyyaml
pdfplumber